v4.19.11 docs: fix indentWithTab prop description (#483)
* docs: fix indentWithTab prop description

The description for the `indentWithTab` prop in the `ReactCodeMirrorProps` interface was incorrect, as it duplicated the description of the `basicSetup` prop

* docs: fix indentWithTab prop description

This commit updates the indentWithTab prop description in the ReactCodeMirrorProps interface, reflecting the changes made in `core/index.tsx`.

The updated description correctly explains that the prop controls whether pressing the `Tab` key inserts a tab character and indents the text (`true`) or behaves according to the browser's default behavior (`false`). ff629d71c6ff9a924430eb1df3b8c5b2c34f893f